add news fragment and clean code
