+++ doc/news/OSW-646.feature.rst
@@ -1,0 +1,1 @@
+Introduce Observing Conditions Applet in digest page
